# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NetWatch Network Security Monitor API
 * OpenAPI spec version: 0.1.0
 */
import type { AlertSeverity } from "./alertSeverity";

export interface Alert {
  id: number;
  type: string;
  severity: AlertSeverity;
  message: string;
  srcIp: string;
  /** @nullable */
  dstIp?: string | null;
  /** @nullable */
  port?: number | null;
  /** @nullable */
  deviceId?: number | null;
  resolved: boolean;
  /** @nullable */
  resolvedAt?: string | null;
  emailSent?: boolean;
  timestamp: string;
}