# lib/api-zod/src/generated/types/alertInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NetWatch Network Security Monitor API
 * OpenAPI spec version: 0.1.0
 */
import type { AlertInputSeverity } from "./alertInputSeverity";

export interface AlertInput {
  type: string;
  severity: AlertInputSeverity;
  message: string;
  srcIp: string;
  dstIp?: string;
  port?: number;
  deviceId?: number;
}