# lib/api-zod/src/generated/types/device.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NetWatch Network Security Monitor API
 * OpenAPI spec version: 0.1.0
 */

export interface Device {
  id: number;
  ip: string;
  /** @nullable */
  mac: string | null;
  label: string;
  /** @nullable */
  vendor?: string | null;
  firstSeen: string;
  lastSeen: string;
  totalBytes: number;
  downloadBytes?: number;
  uploadBytes?: number;
  isBlocked: boolean;
  isOnline?: boolean;
}