# lib/api-zod/src/generated/types/geoTrafficEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NetWatch Network Security Monitor API
 * OpenAPI spec version: 0.1.0
 */

export interface GeoTrafficEvent {
  id: number;
  srcIp: string;
  dstIp: string;
  protocol: string;
  bytes: number;
  direction: string;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lon?: number | null;
  isSuspicious?: boolean;
  timestamp: string;
}