# lib/api-zod/src/generated/types/portScanResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NetWatch Network Security Monitor API
 * OpenAPI spec version: 0.1.0
 */
import type { PortScanResultState } from "./portScanResultState";

export interface PortScanResult {
  ip: string;
  port: number;
  protocol: string;
  state: PortScanResultState;
  /** @nullable */
  service: string | null;
  timestamp: string;
}