# lib/api-zod/src/generated/types/settingsUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NetWatch Network Security Monitor API
 * OpenAPI spec version: 0.1.0
 */

export interface SettingsUpdate {
  emailAlertsEnabled?: boolean;
  alertEmail?: string;
  routerIp?: string;
  routerUser?: string;
  routerPassword?: string;
  scanInterval?: number;
  suspicionThreshold?: number;
  portScanDetection?: boolean;
  bruteForceDetection?: boolean;
  highBandwidthThresholdMb?: number;
}