# lib/api-zod/src/generated/types/testEmailResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NetWatch Network Security Monitor API
 * OpenAPI spec version: 0.1.0
 */

export interface TestEmailResult {
  success: boolean;
  message: string;
}