# lib/api-zod/src/generated/types/trafficEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NetWatch Network Security Monitor API
 * OpenAPI spec version: 0.1.0
 */
import type { TrafficEventDirection } from "./trafficEventDirection";

export interface TrafficEvent {
  id: number;
  srcIp: string;
  dstIp: string;
  srcPort: number;
  dstPort: number;
  protocol: string;
  bytes: number;
  direction: TrafficEventDirection;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lon?: number | null;
  /** @nullable */
  deviceId?: number | null;
  isSuspicious?: boolean;
  /** @nullable */
  suspicionReason?: string | null;
  timestamp: string;
}