# lib/api-zod/src/generated/types/trafficEventInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NetWatch Network Security Monitor API
 * OpenAPI spec version: 0.1.0
 */
import type { TrafficEventInputDirection } from "./trafficEventInputDirection";

export interface TrafficEventInput {
  srcIp: string;
  dstIp: string;
  srcPort: number;
  dstPort: number;
  protocol: string;
  bytes: number;
  direction: TrafficEventInputDirection;
  country?: string;
  city?: string;
  lat?: number;
  lon?: number;
  deviceId?: number;
  isSuspicious?: boolean;
  suspicionReason?: string;
}